+++ src/source/raster_dem_tile_source.js
@@ -8,6 +8,7 @@ import window from '../util/window.js';
 import offscreenCanvasSupported from '../util/offscreen_canvas_supported.js';
 import {OverscaledTileID} from './tile_id.js';
 import RasterTileSource from './raster_tile_source.js';
+import {preloadTiles} from './preload_tiles.js';
 // ensure DEMData is registered for worker transfer on main thread:
 import '../data/dem_data.js';
 
@@ -16,6 +17,9 @@ import type Dispatcher from '../util/dispatcher.js';
 import type Tile from './tile.js';
 import type {Callback} from '../types/callback.js';
 import type {RasterDEMSourceSpecification} from '../style-spec/types.js';
+import type {CameraOptions} from '../ui/camera.js';
+import type {LngLatBoundsLike} from '../geo/lng_lat_bounds.js';
+import type {TilesPreloadProgress} from './preload_tiles.js';
 
 class RasterDEMTileSource extends RasterTileSource implements Source {
     encoding: "mapbox" | "terrarium";
@@ -126,6 +130,32 @@ class RasterDEMTileSource extends RasterTileSource implements Source {
         delete tile.neighboringTiles;
 
         tile.state = 'unloaded';
+    }
+
+    /**
+     * Preloads tiles in the requested viewport.
+     *
+     * @param {LngLatBoundsLike} bounds Center these bounds in the viewport and use the highest
+     *      zoom level up to and including `Map#getMaxZoom()` that fits them in the viewport.
+     * @param {Object} [options] Options supports all properties from {@link CameraOptions}.
+     * @param {Function} [callback] Called when each of the requested tiles is ready or errored.
+     * @returns {number} Number of tiles to load.
+     * @example
+     * map.addSource('mapbox-dem', {
+     *     'type': 'raster-dem',
+     *     'url': 'mapbox://mapbox.mapbox-terrain-dem-v1',
+     *     'tileSize': 512,
+     *     'maxzoom': 14
+     * });
+     *
+     * map.getSource('mapbox-dem').preloadTiles(bbox, {padding: 20}, ({requested, pending, completed, errored}) => {
+     *     if (pending === 0) {
+     *         map.fitBounds(bbox, {duration:0});
+     *     }
+     * });
+     */
+    preloadTiles(bounds: LngLatBoundsLike, options?: CameraOptions, callback?: (progress: TilesPreloadProgress) => void): number {
+        return preloadTiles.call(this, bounds, options, callback);
     }
 
 }
