+++ src/source/vector_tile_source.js
@@ -24,7 +24,7 @@ import type Actor from '../util/actor.js';
 import type {LoadVectorTileResult} from './vector_tile_worker_source.js';
 import type {CameraOptions} from '../ui/camera.js';
 import type {LngLatBoundsLike} from '../geo/lng_lat_bounds.js';
-import type {TilesPreloadProgress} from './source_cache.js';
+import type {TilesPreloadProgress} from './preload_tiles.js';
 
 /**
  * A source containing vector tiles in [Mapbox Vector Tile format](https://docs.mapbox.com/vector-tiles/reference/).
@@ -323,12 +323,12 @@ class VectorTileSource extends Evented implements Source {
      * @param {Function} [callback] Called when each of the requested tiles is ready or errored.
      * @returns {number} Number of tiles to load.
      * @example
-     * map.addSource('some id', {
+     * map.addSource('id', {
      *     type: 'vector',
      *     url: 'mapbox://mapbox.mapbox-streets-v8'
      * });
      *
-     * map.getSource('some id').preloadTiles(bbox, {padding: 20}, ({pending}) => {
+     * map.getSource('id').preloadTiles(bbox, {padding: 20}, ({requested, pending, completed, errored}) => {
      *     if (pending === 0) {
      *         map.fitBounds(bbox, {duration:0});
      *     }
